requirements.txt
langchain
groq
langchain-groq
chromadb
beautifulsoup4
requests
pdfplumber
PyPDF2
tqdm
sentence-transformers
transformers
faiss-cpu
python-dotenv
streamlit
langchain-text-splitters





